requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==22.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django
django-admin-interface==0.26.0
django-auditlog==2.3.0
django-colorfield==0.8.0
django-cors-headers==4.0.0
django-extensions
django-filter
django-model-utils
django-taggit==4.0.0
djangorestframework
djangorestframework-jsonapi
drf-yasg==1.21.5
filelock==3.12.0
flake8==6.0.0
Flask==2.3.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.24.3
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pipenv==2023.6.2
platformdirs==3.5.1
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
virtualenv==20.23.0
virtualenv-clone==0.5.7
Werkzeug==2.3.4
